fix: removed repeated functions
