+++ .changeset/red-chairs-mate.md
@@ -1,0 +1,6 @@
+---
+"@polkadex/polkapool": major
+"@polkadex/numericals": minor
+---
+
+fix: parseScientific function
