+++ apps/polkapool/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@polkadex/blockchain-api": "^1.1.61",
+    "@polkadex/numericals": "^0.3.0",
     "@polkadex/polkadex-api": "^3.1.2",
     "@polkadex/react-providers": "^2.0.0",
     "@polkadex/types": "^1.2.3",
