+++ apps/polkapool/src/constants/index.ts
@@ -1,3 +1,1 @@
 export * from "./assets";
-export * from "./trimFloat";
-export * from "./parseScientific";
