+++ apps/polkapool/src/helpers/index.ts
@@ -1,4 +1,3 @@
-export * from "./truncateString";
 export * from "./typeofChildren";
 export * from "./currentTimestamp";
 export * from "./signAndSend";
